@@ -15,13 +15,13 @@
         "source": "https://github.com/hyperf/hyperf"
     },
     "require": {
-        "php": ">=7.2",
+        "php": ">=8.0",
         "doctrine/instantiator": "^1.2.0",
-        "hyperf/contract": "~2.2.0",
-        "hyperf/utils": "~2.2.0",
-        "hyperf/process": "~2.2.0",
-        "hyperf/pool": "~2.2.0",
-        "php-amqplib/php-amqplib": "^3.0",
+        "hyperf/contract": "~3.0.0",
+        "hyperf/utils": "~3.0.0",
+        "hyperf/process": "~3.0.0",
+        "hyperf/pool": "~3.0.0",
+        "php-amqplib/php-amqplib": "^3.1",
         "psr/container": "^1.0|^2.0",
         "psr/event-dispatcher": "^1.0",
         "psr/log": "^1.0|^2.0|^3.0"
@@ -45,7 +45,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.2-dev"
+            "dev-master": "3.0-dev"
         },
         "hyperf": {
             "config": "Hyperf\\Amqp\\ConfigProvider"
