@@ -31,50 +31,23 @@ class AMQPConnection extends AbstractConnection
 
     public const CONFIRM_CHANNEL_POOL_LENGTH = 10000;
 
-    /**
-     * @var Channel
-     */
-    protected $pool;
+    protected Channel $pool;
 
-    /**
-     * @var Channel
-     */
-    protected $confirmPool;
+    protected Channel $confirmPool;
 
-    /**
-     * @var null|LoggerInterface
-     */
-    protected $logger;
+    protected ?LoggerInterface $logger = null;
 
-    /**
-     * @var int
-     */
-    protected $lastChannelId = 0;
+    protected int $lastChannelId = 0;
 
-    /**
-     * @var null|Params
-     */
-    protected $params;
+    protected ?Params $params = null;
 
-    /**
-     * @var bool
-     */
-    protected $loop = false;
+    protected bool $loop = false;
 
-    /**
-     * @var bool
-     */
-    protected $enableHeartbeat = false;
+    protected bool $enableHeartbeat = false;
 
-    /**
-     * @var ChannelManager
-     */
-    protected $channelManager;
+    protected ChannelManager $channelManager;
 
-    /**
-     * @var bool
-     */
-    protected $exited = false;
+    protected bool $exited = false;
 
     public function __construct(
         string $user,
@@ -105,20 +78,14 @@ class AMQPConnection extends AbstractConnection
         parent::write($data);
     }
 
-    /**
-     * @return static
-     */
-    public function setLogger(?LoggerInterface $logger)
+    public function setLogger(?LoggerInterface $logger): static
     {
         $this->logger = $logger;
 
         return $this;
     }
 
-    /**
-     * @return static
-     */
-    public function setParams(Params $params)
+    public function setParams(Params $params): static
     {
         $this->params = $params;
         return $this;
@@ -222,12 +189,12 @@ class AMQPConnection extends AbstractConnection
         Coroutine::create(function () {
             try {
                 while (true) {
-                    [$frame_type, $channel, $payload] = $this->wait_frame(0);
+                    [$frame_type, $channel, $payload] = $this->wait_frame();
                     $this->channelManager->get($channel)->push([$frame_type, $payload], 0.001);
                 }
             } catch (\Throwable $exception) {
                 $level = $this->exited ? 'warning' : 'error';
-                $this->logger && $this->logger->log($level, 'Recv loop broken. The reason is ' . (string) $exception);
+                $this->logger && $this->logger->log($level, 'Recv loop broken. The reason is ' . $exception);
             } finally {
                 $this->loop = false;
                 if (! $this->exited) {
