@@ -14,20 +14,10 @@ namespace Hyperf\Amqp\Annotation;
 use Attribute;
 use Hyperf\Di\Annotation\AbstractAnnotation;
 
-/**
- * @Annotation
- * @Target({"CLASS"})
- */
 #[Attribute(Attribute::TARGET_CLASS)]
 class Producer extends AbstractAnnotation
 {
-    /**
-     * @var string
-     */
-    public $exchange = '';
+    public string $exchange = '';
 
-    /**
-     * @var string
-     */
-    public $routingKey = '';
+    public string $routingKey = '';
 }
