@@ -16,15 +16,9 @@ use Throwable;
 
 class FailToConsume extends ConsumeEvent
 {
-    /**
-     * @var Throwable
-     */
-    protected $throwable;
-
-    public function __construct(ConsumerMessageInterface $message, Throwable $throwable)
+    public function __construct(ConsumerMessageInterface $message, protected Throwable $throwable)
     {
         parent::__construct($message);
-        $this->throwable = $throwable;
     }
 
     public function getThrowable(): Throwable
