@@ -22,14 +22,8 @@ use Psr\Container\ContainerInterface;
  */
 class BeforeMainServerStartListener implements ListenerInterface
 {
-    /**
-     * @var ContainerInterface
-     */
-    private $container;
-
-    public function __construct(ContainerInterface $container)
+    public function __construct(private ContainerInterface $container)
     {
-        $this->container = $container;
     }
 
     /**
