@@ -17,20 +17,11 @@ use Hyperf\Utils\ApplicationContext;
 
 abstract class ProducerMessage extends Message implements ProducerMessageInterface
 {
-    /**
-     * @var string
-     */
-    protected $payload = '';
+    protected mixed $payload = '';
 
-    /**
-     * @var string
-     */
-    protected $routingKey = '';
+    protected array|string $routingKey = '';
 
-    /**
-     * @var array
-     */
-    protected $properties
+    protected array $properties
         = [
             'content_type' => 'text/plain',
             'delivery_mode' => Constants::DELIVERY_MODE_PERSISTENT,
