@@ -22,15 +22,8 @@ class RpcClient extends Builder
 {
     protected $poolChannels = [];
 
-    /**
-     * @var int
-     */
-    protected $maxChannels;
-
-    public function __construct(ContainerInterface $container, ConnectionFactory $factory, int $maxChannels = 64)
+    public function __construct(ContainerInterface $container, ConnectionFactory $factory, protected int $maxChannels = 64)
     {
-        $this->maxChannels = $maxChannels;
-
         parent::__construct($container, $factory);
     }
 
