@@ -16,14 +16,14 @@ use Hyperf\Amqp\Result;
 
 class DemoConsumer extends ConsumerMessage
 {
-    protected $exchange = 'hyperf';
+    protected string $exchange = 'hyperf';
 
-    protected $routingKey = [
+    protected array|string $routingKey = [
         'hyperf1',
         'hyperf2',
     ];
 
-    protected $queue = 'hyperf';
+    protected ?string $queue = 'hyperf';
 
     public function consume($data): string
     {
