@@ -13,6 +13,7 @@ namespace HyperfTest\Amqp\Stub;
 
 use Hyperf\Amqp\ConnectionFactory;
 use Hyperf\Amqp\Consumer;
+use Hyperf\Amqp\IO\IOFactory;
 use Hyperf\Amqp\Pool\PoolFactory;
 use Hyperf\Config\Config;
 use Hyperf\Contract\ConfigInterface;
@@ -68,7 +69,7 @@ class ContainerStub
             return new Consumer($container, $container->get(ConnectionFactory::class), $container->get(StdoutLoggerInterface::class));
         });
         $container->shouldReceive('get')->with(FormatterInterface::class)->andReturn(new DefaultFormatter());
-
+        $container->shouldReceive('get')->with(IOFactory::class)->andReturn(new IOFactory());
         return $container;
     }
 
